Update PrivateKey(string) constructor to take 32-byte / 64 character key
The previous PrivateKey(string) constructor expects an extended private key string instead of a 32-byte / 64-character string. The updates now constraint the constructor to take a 64-character string as a private key. This string is then used as a seed to derive the extended private key used by `Chaos.NaCL`. See the getter and setter for `KeyBytes`.